buff azalea + nerf fern spawn

+++ Client/Renderer/RenderMob.c
@@ -30,7 +30,7 @@ void rr_component_mob_render(EntityIdx entity, struct rr_game *game)
     }
     if (physical->animation > 2 * M_PI)
         physical->animation = fmod(physical->animation, 2 * M_PI);
-    float sinusoid_animation = sin(physical->animation);
+    float sinusoid_animation = sinf(physical->animation);
 
     rr_renderer_render_mob(renderer, game, mob->id, sinusoid_animation,
                            physical->turning_animation - physical->lerp_angle);
